+++ R/conditions.R
@@ -13,9 +13,7 @@ arg_name <- function(quos, index) {
 # Common ------------------------------------------------------------------
 
 stop_eval_tidy <- function(e, index, dots, fn) {
-  data  <- peek_mask()$full_data()
-  group <- peek_mask()$get_current_group()
-
+  data  <- context_get("..mask")$full_data()
   expr  <- as_label(quo_get_expr(dots[[index]]))
   name  <- arg_name(dots, index)
 
@@ -23,7 +21,7 @@ stop_eval_tidy <- function(e, index, dots, fn) {
     "`{fn}()` argument `{name}` errored.",
     i = "`{name}` is {expr}",
     x = conditionMessage(e),
-    i = if(is_grouped_df(data)) "The error occured in group {group}."
+    i = if(is_grouped_df(data)) "The error occured in group {cur_group_id()}."
   ))
 }
 
@@ -100,8 +98,7 @@ stop_summarise_unsupported_type <- function(result, index, dots) {
     abort(class = "dplyr_summarise_unsupported_type", result = result)
   }
 
-  data  <- peek_mask()$full_data()
-  group <- peek_mask()$get_current_group()
+  data  <- context_get("..mask")$full_data()
   expr  <- as_label(quo_get_expr(dots[[index]]))
   name  <- arg_name(dots, index)
 
@@ -109,7 +106,7 @@ stop_summarise_unsupported_type <- function(result, index, dots) {
   abort(glue_c(
     "`summarise()` argument `{name}` must be a vector.",
     i = "`{name}` is {expr}",
-    i = if(is_grouped_df(data)) "The error occured in group {group}.",
+    i = if(is_grouped_df(data)) "The error occured in group {cur_group_id()}.",
     x = "Result should be a vector, not {as_friendly_type(typeof(result))}."
   ))
 
@@ -142,28 +139,26 @@ stop_mutate_not_vector <- function(result, index, dots) {
 
   name <- arg_name(dots, index)
   expr <- as_label(quo_get_expr(dots[[index]]))
-  data <- peek_mask()$full_data()
-  group <- peek_mask()$get_current_group()
+  data <- context_get("..mask")$full_data()
 
   abort(glue_c(
     "`mutate()` argument `{name}` must be a vector.",
     i = "`{name}` is {expr}.",
     x = "Result should be a vector, not {as_friendly_type(typeof(result))}.",
-    i = if(is_grouped_df(data)) "The error occured in group {group}."
+    i = if(is_grouped_df(data)) "The error occured in group {cur_group_id()}."
   ))
 }
 
 stop_mutate_recycle_incompatible_size <- function(cnd, index, dots) {
   name <- arg_name(dots, index)
   expr <- as_label(quo_get_expr(dots[[index]]))
-  data <- peek_mask()$full_data()
-  group <- peek_mask()$get_current_group()
+  data <- context_get("..mask")$full_data()
 
   abort(glue_c(
     "`mutate()` argument `{name}` must be recyclable.",
     i = "`{name}` is {expr}",
     x = conditionMessage(cnd),
-    i = if(is_grouped_df(data)) "The error occured in group {group}."
+    i = if(is_grouped_df(data)) "The error occured in group {cur_group_id()}."
   ))
 }
 
@@ -173,7 +168,7 @@ stop_summarise_incompatible_size <- function(size, group, index, expected_sizes,
     abort(class = "dplyr_summarise_incompatible_size", size = size, group = group)
   }
 
-  data <- peek_mask()$full_data()
+  data <- context_get("..mask")$full_data()
   name <- arg_name(dots, index)
   expr <- as_label(quo_get_expr(dots[[index]]))