+++ R/context.R
@@ -6,14 +6,27 @@
 #' `mutate()`
 #'
 #' * `n()` gives the current group size.
+#' * `cur_data()` gives the current data for the current group (exclusing
+#'   grouping variables)
 #' * `cur_group()` gives the group keys, a tibble with one row and one column
 #'   for each grouping variable.
 #' * `cur_group_id()` gives a unique numeric identifier for the current group.
 #' * `cur_column()` gives the current column (in [across()] only).
 #'
+#' See [group_data()] for equivalent functions that return values for all
+#' groups.
+#'
+#' @section data.table:
+#' If you're familiar with data.table:
+#'
+#' * `cur_data()` <-> `.SD`
+#' * `cur_group_id()` <-> `.GRP`
+#' * `cur_group()` <-> `.BY`
+#' * `cur_group_rows()` <-> `.I`
+#'
 #' @examples
 #' df <- tibble(
-#'   g = rep(letters[1:3], 1:3),
+#'   g = sample(rep(letters[1:3], 1:3)),
 #'   x = runif(6),
 #'   y = runif(6)
 #' )
@@ -22,6 +35,10 @@
 #' gf %>% summarise(n = n())
 #'
 #' gf %>% mutate(id = cur_group_id())
+#' gf %>% summarise(row = cur_group_rows())
+#' gf %>% summarise(data = list(cur_group()))
+#' gf %>% summarise(data = list(cur_data()))
+#'
 #' gf %>% mutate(across(everything(), ~ paste(cur_column(), round(.x, 2))))
 #' @name context
 NULL
@@ -29,19 +46,33 @@ NULL
 #' @rdname context
 #' @export
 n <- function() {
-  context_get("..group_size")
+  length(context_get("..mask")$current_rows())
+}
+
+#' @rdname context
+#' @export
+cur_data <- function() {
+  mask <- context_get("..mask")
+  data <- mask$full_data()
+  mask$pick(setdiff(names(data), group_vars(data)))
 }
 
 #' @rdname context
 #' @export
 cur_group <- function() {
-  peek_mask()$current_key()
+  context_get("..mask")$current_key()
 }
 
 #' @rdname context
 #' @export
 cur_group_id <- function() {
-  context_get("..group_number")
+  context_get("..mask")$get_current_group()
+}
+
+#' @rdname context
+#' @export
+cur_group_rows <- function() {
+  context_get("..mask")$current_rows()
 }
 
 #' @rdname context