+++ R/tbl-df.r
@@ -4,19 +4,11 @@ poke_mask <- function(mask) {
   old
 }
 
-peek_mask <- function() {
-  context_env[["..mask"]] %||% abort("No dplyr data mask registered")
-}
 
 scoped_mask <- function(mask, frame = caller_env()) {
   old_mask <- poke_mask(mask)
-  old_group_size <- context_env[["..group_size"]]
-  old_group_number <- context_env[["..group_number"]]
-
   expr <- call2(on.exit, expr({
     poke_mask(!!old_mask)
-    context_env[["..group_size"]] <- !!old_group_size
-    context_env[["..group_number"]] <- !!old_group_number
   }), add = TRUE)
   eval_bare(expr, frame)
 }
@@ -25,7 +17,6 @@ DataMask <- R6Class("DataMask",
   public = list(
     initialize = function(data, caller, rows = group_rows(data)) {
       frame <- caller_env(n = 2)
-      tidyselect::scoped_vars(tbl_vars(data), frame)
       scoped_mask(self, frame)
 
       private$rows <- rows
@@ -89,6 +80,10 @@ DataMask <- R6Class("DataMask",
 
     pick = function(vars) {
       eval_tidy(quo(tibble(!!!syms(vars))), private$mask)
+    },
+
+    current_rows = function() {
+      private$rows[[private$current_group]]
     },
 
     current_key = function() {